+++ Lingarr.Core/Enum/TranslationStatus.cs
@@ -7,5 +7,6 @@ public enum TranslationStatus
     Completed,
     Failed,
     Cancelled,
-    Interrupted
+    Interrupted,
+    Partial
 }